refactor(config): break stuff out in prep for reorg
