+++ src/config/oranda_config/analytics.rs
@@ -1,11 +1,9 @@
-use crate::config::Config;
+use crate::config::ApplyLayer;
 use axohtml::elements::script;
 
 use axohtml::{html, unsafe_text};
 use schemars::JsonSchema;
 use serde::Deserialize;
-
-use super::ApplyLayer;
 
 #[derive(Debug, Deserialize, JsonSchema)]
 pub struct GoogleTracking {
@@ -31,13 +29,13 @@ pub struct UnamiTracking {
 
 #[derive(Debug, Deserialize, JsonSchema)]
 #[serde(rename_all = "lowercase")]
-pub enum Analytics {
+pub enum AnalyticsConfig {
     Google(GoogleTracking),
     Plausible(PlausibleTracking),
     Fathom(FathomTracking),
     Unami(UnamiTracking),
 }
-impl ApplyLayer for Analytics {
+impl ApplyLayer for AnalyticsConfig {
     fn apply_layer(&mut self, layer: Self) {
         // FIXME: this is kinda goofy but there's not an obvious thing to do
         // if we need to change the enum variant and we care about preserving things
@@ -57,29 +55,28 @@ impl GoogleTracking {
     }
 }
 
-pub fn snippet(config: &Config) -> Option<Box<script<String>>> {
-    let analytics = config.analytics.as_ref();
-
-    match analytics {
-        None => None,
-        Some(Analytics::Fathom(f)) => {
-            Some(html!(<script defer=true src=FATHOM_SCRIPT_URL data-site=&f.site ></script>))
-        }
-        Some(Analytics::Unami(u)) => Some(
-            html!(<script async=true defer=true src=&u.script_url data-website-id=&u.website></script>),
-        ),
-        Some(Analytics::Google(g)) => {
-            let script_url = format!("{}?id={}", GOOGLE_SCRIPT_URL, g.tracking_id);
-            Some(html!(
-                <script async=true src=&script_url></script>
-            ))
-        }
-        Some(Analytics::Plausible(p)) => {
-            let url = PLAUSIBLE_SCRIPT_URL.to_string();
-            let script_url = p.script_url.as_ref().unwrap_or(&url);
-            Some(html!(
-                <script defer=true data-domain=&p.domain src=script_url></script>
-            ))
+impl AnalyticsConfig {
+    pub fn snippet(&self) -> Box<script<String>> {
+        match self {
+            AnalyticsConfig::Fathom(f) => {
+                html!(<script defer=true src=FATHOM_SCRIPT_URL data-site=&f.site ></script>)
+            }
+            AnalyticsConfig::Unami(u) => {
+                html!(<script async=true defer=true src=&u.script_url data-website-id=&u.website></script>)
+            }
+            AnalyticsConfig::Google(g) => {
+                let script_url = format!("{}?id={}", GOOGLE_SCRIPT_URL, g.tracking_id);
+                html!(
+                    <script async=true src=&script_url></script>
+                )
+            }
+            AnalyticsConfig::Plausible(p) => {
+                let url = PLAUSIBLE_SCRIPT_URL.to_string();
+                let script_url = p.script_url.as_ref().unwrap_or(&url);
+                html!(
+                    <script defer=true data-domain=&p.domain src=script_url></script>
+                )
+            }
         }
     }
 }
