+++ src/config/oranda_config/artifacts.rs
@@ -2,7 +2,7 @@ use indexmap::IndexMap;
 use schemars::JsonSchema;
 use serde::Deserialize;
 
-use super::{ApplyLayer, ApplyOptExt};
+use crate::config::{ApplyLayer, ApplyOptExt};
 
 #[derive(Debug, Deserialize, JsonSchema)]
 #[serde(rename_all = "lowercase")]
@@ -32,14 +32,14 @@ enum ArtifactSystem {
 }
 
 #[derive(Debug, Default, Deserialize, JsonSchema)]
-pub struct Artifacts {
+pub struct ArtifactsConfig {
     #[serde(default)]
     pub cargo_dist: Option<bool>,
     #[serde(default)]
     pub package_managers: Option<IndexMap<String, String>>,
 }
 
-impl ApplyLayer for Artifacts {
+impl ApplyLayer for ArtifactsConfig {
     fn apply_layer(&mut self, layer: Self) {
         self.cargo_dist.apply_opt(layer.cargo_dist);
         // FIXME: should this get merged with e.g. `extend?`
@@ -47,7 +47,7 @@ impl ApplyLayer for Artifacts {
     }
 }
 
-impl Artifacts {
+impl ArtifactsConfig {
     pub fn has_some(&self) -> bool {
         self.cargo_dist() || self.package_managers.is_some()
     }
