+++ src/site/layout/mod.rs
@@ -1,7 +1,8 @@
 use axohtml::{html, text};
 
-use crate::config::{theme, Config};
+use crate::config::Config;
 use crate::errors::*;
+use crate::site::oranda_theme::OrandaTheme;
 
 pub mod css;
 mod footer;
@@ -27,7 +28,7 @@ impl Layout {
     }
 
     pub fn new(config: &Config) -> Result<Self> {
-        let theme = theme::css_class(&config.styles.theme());
+        let theme = OrandaTheme::css_class(&config.styles.theme());
         let name = &config.name;
         let header = match config.no_header {
             true => None,
@@ -61,7 +62,7 @@ impl Layout {
             &config.path_prefix,
             &config.styles.oranda_css_version,
         )?;
-        let analytics = Analytics::new(config)?;
+        let analytics = Analytics::new(&config.analytics)?;
         let template_html: String = html!(
         <html lang="en" id="oranda" class=theme>
             <head>
